pyproject.toml
[tool.poetry]
name = "tank-project"
version = "0.1.12"
description = ""
authors = ["Jana-TT <janatt2056@gmail.com>"]
readme = "README.md"
packages = [{ include = "src", from = "." }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
pydantic = "^2.6.4"
httpx = "^0.27.0"
ipykernel = "^6.29.4"
async-db-polars = { version = "^0.1.7", extras = ["pg"] }

[tool.poetry.group.dev.dependencies]
curlify2 = "^2.0.0"
ruff = "^0.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
main = "src.main:main"
